fix(crc_check): change to do check on 16-bit data

+++ components/crc_check/crc_check.h
@@ -1,5 +1,5 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-unsigned int crc32(uint8_t message);
-bool is_crc_passed(uint8_t message, unsigned int crc_to_check);
+unsigned int crc32(uint16_t message);
+bool is_crc_passed(uint16_t message, unsigned int crc_to_check);
